requirements.txt
spacy
nltk
textblob
langdetect
vaderSentiment
deep-translator
textblob_fr
pandas
numpy
scikit-learn
selenium
webdriver-manager
streamlit
plotly
Pillow
wordcloud
matplotlib
langid